requirements.txt
google-genai
flask
requests